LayoutTests/fast/multicol/balance-short-trailing-empty-block.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<p>Both words below should be in the the right column.</p>
<div id="console"></div>
<div id="mc" style="-webkit-columns:2; columns:2; orphans:1; widows:1; -webkit-column-rule:1px solid; column-rule:1px solid; background:olive;">
    <div style="height:120px;"></div>
    <div style="line-height:80px;">right<br>column</div>
    <div style="height:30px;"></div>
</div>
<script>
    description("Tests that a block a column boundary is recorded properly for balancing");
    shouldBe("mc.offsetHeight", "190");
</script>
