LayoutTests/fast/multicol/balance-short-trailing-empty-block-expected.txt

Both words below should be in the the right column.

Tests that a block a column boundary is recorded properly for balancing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mc.offsetHeight is 190
PASS successfullyParsed is true

TEST COMPLETE
right
column
